# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Taxis",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://cooperativa-taxi.local"
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "operationId": "3265ab79c9b880008ed804d6b5e3059a",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number of the list",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "status filter"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all the users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "We have an error"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create  a new user",
                "operationId": "7cfe4acccb16b697b910fb2c3d8eb350",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create  a new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "We have an error"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get one user",
                "operationId": "3b1806a5ed3afbac8c1d1b073569cc7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show the user information by Id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "We have an error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update  a user",
                "operationId": "d273e38b4ab89cda9bf08737d4bd1dcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update  a user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "We have an error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete  a new user",
                "operationId": "5aa4ff23a2e5a35e34b971342b7ca960",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete a user"
                    },
                    "404": {
                        "description": "We have an error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}/status/{status}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Change the user status",
                "operationId": "1dd4b6fe1e4b29e50efe52b5c4fea239",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id"
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "description": "user id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Change the user status"
                    },
                    "404": {
                        "description": "We have an error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "example": "user@gmail.com"
                    },
                    "password": {
                        "description": "User password",
                        "type": "string"
                    },
                    "status": {
                        "description": "User status",
                        "type": "string",
                        "example": "Active"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        }
    }
}